Add level indicator layout
